feat: implement LoginPageClient component for improved login handling and redirection
